chore: add pyproject and fix requirements
