+++ rag_ed/agents/__init__.py
@@ -1,0 +1,3 @@
+"""Agent demo modules for rag_ed."""
+
+__all__ = []
