+++ rag_ed/loaders/__init__.py
@@ -1,0 +1,3 @@
+"""Loaders for educational content."""
+
+__all__ = []
