+++ requirements.txt
@@ -2,7 +2,6 @@ black
 langchain-core
 langchain-community
 unstructured
-pdfminer
 pdfminer.six==20221105
 pi_heif
 unstructured-inference
@@ -11,4 +10,4 @@ pdf2image
 unstructured-pytesseract
 jq
 smolagents
-langchain_openai
+langchain-openai
